Remove unnecesary context instance from broker
On creation the broker was setting a context, this would have allowed
the broker to handle messages without panicking before Start was called,
but that would hide incorrect usage of broker.

Now if broker tires to hanlde messages before start is called it will
most likly panic due to a nil context.