Fix race conditions in broker
The fine grained locking made it hard to understand what was being
locked and what was not being locked. This had resulted in race
conditions being introduced into the broker. The large scale locking
should result in safe execution since all exported methods are locked
over and no goroutines are created by the broker.

There was a race condition in validateTiming because validateTiming (see
below) treated the inbox and latest layer as if they were a single
atomically set value.

In fact both the inbox and the latest layer were being set individually
and additionally validateTiming lacked any synchronisation with the 2
values.

This led to a situation where no inbox could be found for the layer
referenced by the message, but then finding that the latest layer
matched the message's layer, resulting in errRegistration being
returned.

This is fixed by adding a new mutex that ensures that validateTiming
sees any updates to these two values atomically.

With this fix the outcome will instead be, either that the message is
treated as an early message, or that the inbox would be found and the
message delivered as usual.

func (b *Broker) validateTiming(ctx context.Context, m *Message) error {
        msgInstID := m.Layer
        if b.getInbox(msgInstID) != nil {
                b.Log.WithContext(ctx).With().Debug("instance exists", msgInstID)
                return nil
        }

        latestLayer := b.getLatestLayer()

        if msgInstID.Before(latestLayer) {
                return fmt.Errorf("%w: msg %v, latest %v", errUnregistered, msgInstID.Uint32(), latestLayer)
        }

        // current layer
        if msgInstID == latestLayer {
                return fmt.Errorf("%w: latest %v", errRegistration, latestLayer)
        }

        // early msg
        if msgInstID == latestLayer.Add(1) {
                return fmt.Errorf("%w: latest %v", errEarlyMsg, latestLayer)
        }

There was also another race condition where early messages could be added to
pending after a consensus process had registered for that layer,
essentially resulting in the early messages being lost.