Remove pending messages map from broker
Currently the broker holds a lock across the whole of handleMessage,
although this fixes the race conditions that the broker had it also
means that if one outbox channel gets full the broker will not be able
to service any other outbox channels. In order to solve this we need to
avoid locking over the send on outbox chanels, but that is currently
hard to do while maintaining safety.

In order to be able to simplify locking within handleMessage removed the
pending messages map which removed the need to lock over it.

Alos removed validateTiming to avoid doubling the logic to retrieve and
handle the outbox. Also having to interpret the returned error to
understand if a message was early overly complicated what could be a
simple check.